Skip Scraper is done
There is a weird bug on skip for Dairy Queen.

See Item is the price text, and the see item function to see how expensive it is doesn't even work on skip's website!! so how can we see the price at all? We skip the "See item" stuff for now