Fixed a bug that's found in all scrapers
Basically, if you imported the respective scrape functions to another file, a web driver would spawn without you calling those functions.

What's happening is that because there is code outside of those functions, that code is being executed. In some scrapers, a web driver can be found outside of functions, so that's being executed without the functions themselves being run.